Split app.js phase 4c-2: extract onboarding module
In-app tests pass (132/132).
