+++ src/state.js
@@ -82,6 +82,7 @@ export const state = {
     lastMigrationHeaderCopyKey: '',
     lastMigrationHowtoCopyKey: '',
     extensionSetupPausedForBackNavigation: false,
+    forceShowEulaThisSession: false,
 };
 
 // Alias for use inside functions whose local parameter is also named `state`
